requirements.txt
# FastAPI and Server Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Machine Learning and Image Processing - Updated for Python 3.13 compatibility
tensorflow>=2.15.0
opencv-python>=4.8.0
Pillow>=10.1.0
numpy>=1.26.0

# ONNX Conversion Dependencies - Now compatible with Python 3.13
onnx>=1.17.0
tf2onnx>=1.16.0
onnxruntime>=1.16.0
flatbuffers>=24.3.25

# Data Validation and Serialization
pydantic>=2.5.0

# Logging and Utilities
python-json-logger>=2.0.0

# Development and Testing (Optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Environment and Configuration
python-dotenv>=1.0.0

# CORS and Middleware
starlette>=0.27.0
